Created manage.py and moved testsettings to testapp #364
