+++ docs/develop/testing.rst
@@ -34,6 +34,10 @@ or inside a project with ``guardian`` set at ``INSTALLED_APPS``::
 
     $ python manage.py test guardian
 
+or using the bundled ``testapp`` project::
+
+    $ cd guardian/testapp/
+    $ python manage.py test
 
 Coverage support
 ----------------
